# [chore] [exporter/splunkhec] Reuse jsoniter stream/buffer