--- a/exporter/splunkhecexporter/testdata/hec_log_event.json
+++ b/exporter/splunkhecexporter/testdata/hec_log_event.json
@@ -1,0 +1,11 @@
+{
+  "host": "myhost",
+  "source": "myapp",
+  "sourcetype": "myapp-type",
+  "index": "myindex",
+  "event": "mylog",
+  "fields": {
+    "otel.log.name": "0_0_0",
+    "custom": "custom"
+  }
+}