--- a/exporter/splunkhecexporter/testdata/hec_metric_event.json
+++ b/exporter/splunkhecexporter/testdata/hec_metric_event.json
@@ -1,0 +1,14 @@
+{
+  "host": "unknown",
+  "event": "metric",
+  "fields": {
+    "k/n1": "vn1",
+    "k/r0": "vr0",
+    "k/r1": "vr1",
+    "metric_name:gauge_double_with_dims": 1234.5678,
+    "metric_type": "Gauge",
+    "k0": "v0",
+    "k1": "v1",
+    "k/n0": "vn0"
+  }
+}