--- a/exporter/splunkhecexporter/testdata/hec_span_event.json
+++ b/exporter/splunkhecexporter/testdata/hec_span_event.json
@@ -1,0 +1,20 @@
+{
+  "time": 1,
+  "host": "unknown",
+  "event": {
+    "trace_id": "01010101010101010101010101010101",
+    "span_id": "0000000000000001",
+    "parent_span_id": "0102030405060708",
+    "name": "root",
+    "end_time": 2000000000,
+    "kind": "SPAN_KIND_UNSPECIFIED",
+    "status": {
+      "message": "ok",
+      "code": "STATUS_CODE_OK"
+    },
+    "start_time": 1000000000
+  },
+  "fields": {
+    "resource": "R1"
+  }
+}